+++ balancer-js/src/lib/constants/pools.ts
@@ -644,6 +644,7 @@ const POOLS_MAP = {
   [Network.ARBITRUM]: POOLS_ARBITRUM,
 };
 
-export function POOLS(networkId: keyof typeof POOLS_MAP): Pools {
-  return POOLS_MAP[networkId] ? POOLS_MAP[networkId] : POOLS_GENERIC;
+export function POOLS(networkId: Network): Pools {
+  const id = networkId as keyof typeof POOLS_MAP;
+  return POOLS_MAP[id] ? POOLS_MAP[id] : POOLS_GENERIC;
 }
