+++ balancer-js/src/modules/data/gauge-shares/types.ts
@@ -5,12 +5,6 @@ export enum GaugeShareAttributes {
   Balance = 'balance',
 }
 
-export type GaugeShareAttribute =
-  | GaugeShareAttributes.Id
-  | GaugeShareAttributes.UserAddress
-  | GaugeShareAttributes.GaugeId
-  | GaugeShareAttributes.Balance;
-
 export interface GaugeShare {
   id: string;
   balance: string;
