+++ balancer-js/src/modules/data/pool-shares/types.ts
@@ -5,12 +5,6 @@ export enum PoolShareAttributes {
   UserAddress = 'userAddress',
 }
 
-export type PoolShareAttribute =
-  | PoolShareAttributes.Id
-  | PoolShareAttributes.PoolId
-  | PoolShareAttributes.UserAddress
-  | PoolShareAttributes.Balance;
-
 export interface PoolShare {
   id: string;
   userAddress: string;
